+++ src/addons/display_commander/settings/advanced_tab_settings.cpp
@@ -70,7 +70,9 @@ AdvancedTabSettings::AdvancedTabSettings()
       d3d11_texture_caching_1d_enabled("D3D11TextureCaching1DEnabled", false, "DisplayCommander"),
       d3d11_texture_caching_3d_enabled("D3D11TextureCaching3DEnabled", false, "DisplayCommander"),
       texture_cache_content_hash_cap_kb("TextureCacheContentHashCapKb", 64, 1, 1048576, "DisplayCommander"),
-      dump_textures_enabled("DumpTexturesEnabled", false, "DisplayCommander") {}
+      dump_textures_enabled("DumpTexturesEnabled", false, "DisplayCommander"),
+      show_steam_achievement_notifications("ShowSteamAchievementNotifications", false, "DisplayCommander"),
+      show_steam_achievement_counter_increased("ShowSteamAchievementCounterIncreased", false, "DisplayCommander") {}
 
 void AdvancedTabSettings::LoadAll() {
     // Get all settings for smart logging
@@ -109,6 +111,8 @@ void AdvancedTabSettings::SaveAll() {
     d3d11_texture_caching_3d_enabled.Save();
     texture_cache_content_hash_cap_kb.Save();
     dump_textures_enabled.Save();
+    show_steam_achievement_notifications.Save();
+    show_steam_achievement_counter_increased.Save();
 }
 
 std::vector<ui::new_ui::SettingBase*> AdvancedTabSettings::GetAllSettings() {
@@ -129,7 +133,7 @@ std::vector<ui::new_ui::SettingBase*> AdvancedTabSettings::GetAllSettings() {
         &presentmon_provider_dxgkrnl, &presentmon_provider_dxgi, &presentmon_provider_dwm, &presentmon_provider_d3d9,
         &disable_dpi_scaling, &enable_dx11_vtable_hooks, &texture_tracking_enabled, &d3d11_texture_caching_enabled,
         &d3d11_texture_caching_1d_enabled, &d3d11_texture_caching_3d_enabled, &texture_cache_content_hash_cap_kb,
-        &dump_textures_enabled};
+        &dump_textures_enabled, &show_steam_achievement_notifications, &show_steam_achievement_counter_increased};
 }
 
 }  // namespace settings